# Conflicts:
#	backend/eln/eln-api/src/main/java/com/epam/indigoeln/eln/model/RevisionDetailsDTO.java
#	backend/eln/eln-core/src/main/java/com/epam/indigoeln/eln/mapper/ExperimentMapper.java
#	backend/eln/eln-core/src/main/java/com/epam/indigoeln/eln/mapper/NotebookMapper.java
#	backend/eln/eln-core/src/main/java/com/epam/indigoeln/eln/mapper/ProjectMapper.java
#	backend/eln/eln-core/src/main/java/com/epam/indigoeln/eln/service/RevisionService.java
#	backend/eln/eln-core/src/main/java/com/epam/indigoeln/eln/util/PatchUtil.java
#	backend/eln/eln-core/src/test/java/com/epam/indigoeln/reaction/util/CalculationReportBuilder.java
#	deployment-aws/src/main/java/com/epam/indigoeln/aws/CloudFrontStack.java
#	indigo-frontend/src/app/pages/experiment/experiment-info/experiment-info.component.ts
#	indigo-frontend/src/app/pages/experiment/experiment-versions/experiment-versions.component.ts
#	indigo-frontend/src/app/pages/experiment/sample-search/sample-search.component.ts
#	indigo-frontend/src/app/pages/experiment/stoichiometry/product-batch-summary-table/product-batch-summary-table.component.ts
#	indigo-frontend/src/app/pages/experiment/stoichiometry/reaction-inputs-table/reaction-inputs-table.component.ts
#	indigo-frontend/src/app/pages/experiment/stoichiometry/reaction-scheme-view/reaction-scheme-view.component.ts
#	indigo-frontend/src/app/pages/experiment/stoichiometry/reaction-view/reaction-view.component.ts
#	indigo-frontend/src/app/pages/search/global-search/global-search.component.ts
#	indigo-frontend/src/core/services/experiment/experiment-detail.service.ts
#	indigo-frontend/src/core/services/experiment/mutation-builder.service.ts
#	indigo-frontend/src/core/types/entities/compound.i.ts
#	indigo-frontend/src/core/types/entities/experiments/experiment-detail.i.ts
#	indigo-frontend/src/core/types/entities/experiments/experiment.i.ts
#	indigo-frontend/src/core/types/entities/values.i.ts

+++ backend/eln/eln-api/src/main/java/com/epam/indigoeln/compound/model/SampleRegistrationRequest.java
@@ -14,6 +14,7 @@ import lombok.Data;
 import lombok.With;
 import org.jspecify.annotations.Nullable;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -34,7 +35,7 @@ public class SampleRegistrationRequest {
     private EnteredValue<MolarityUnit> molarity;
 
     @Nullable
-    private Double purity;
+    private BigDecimal purity;
 
     @Nullable
     @Size(min = 1)
